+++ FMS.Business/Classes/DataObjects/FleetManagement/usp_GetRunListingReport.vb
@@ -4,7 +4,7 @@
         Public Property RunNUmber As System.Nullable(Of Integer)
         Public Property RunNum As String
         Public Property RunDescription As String
-        Public Property RunDriver As System.Nullable(Of Short)
+        Public Property RunDriver As System.Nullable(Of Integer)
         Public Property DriverName As String
         Public Property MondayRun As Boolean
         Public Property TuesdayRun As Boolean
